feat: implement data validation and fix handlers

+++ internal/handlers/user/handler.go
@@ -2,6 +2,7 @@ package userHandler
 
 import (
 	"github.com/gofiber/fiber/v2"
+	"github.com/google/uuid"
 	"github.com/jon-mun/learn-go-fiber/database"
 	"github.com/jon-mun/learn-go-fiber/internal/model"
 )
@@ -18,8 +19,17 @@ func GetUsers(c *fiber.Ctx) error {
 		})
 	}
 
+	var responseData []interface{}
+	for _, user := range users {
+		responseData = append(responseData, fiber.Map{
+			"id":    user.ID,
+			"name":  user.Name,
+			"email": user.Email,
+		})
+	}
+
 	return c.JSON(fiber.Map{
-		"data": users,
+		"data": responseData,
 	})
 }
 
@@ -33,19 +43,45 @@ func CreateUser(c *fiber.Ctx) error {
 		})
 	}
 
-	db.Create(&user)
+	id, err := uuid.NewUUID()
+
+	if err != nil {
+		return c.Status(fiber.StatusInternalServerError).JSON(fiber.Map{
+			"message": "Cannot generate UUID",
+		})
+	}
+
+	user.ID = id
+	result := db.Create(&user)
+
+	if result.Error != nil {
+		return c.Status(fiber.StatusInternalServerError).JSON(fiber.Map{
+			"message": result.Error.Error(),
+		})
+	}
 
 	return c.JSON(fiber.Map{
-		"data": user,
+		"data": fiber.Map{
+			"id":    user.ID,
+			"name":  user.Name,
+			"email": user.Email,
+		},
 	})
 }
 
 func GetUser(c *fiber.Ctx) error {
 	id := c.Params("id")
+
 	db := database.DB
 	var user model.User
 
-	db.Find(&user, id)
+	// Convert the string ID to a UUID
+	parsedID, err := uuid.Parse(id)
+	if err != nil {
+		return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{"error": "Invalid UUID format"})
+	}
+
+	db.Find(&user, parsedID)
 
 	return c.JSON(fiber.Map{
 		"data": user,
@@ -54,10 +90,16 @@ func GetUser(c *fiber.Ctx) error {
 
 func UpdateUser(c *fiber.Ctx) error {
 	id := c.Params("id")
+
 	db := database.DB
 	var user model.User
 
-	db.Find(&user, id)
+	parsedID, err := uuid.Parse(id)
+	if err != nil {
+		return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{"error": "Invalid UUID format"})
+	}
+
+	db.Find(&user, parsedID)
 
 	if err := c.BodyParser(&user); err != nil {
 		return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{
@@ -65,10 +107,20 @@ func UpdateUser(c *fiber.Ctx) error {
 		})
 	}
 
-	db.Save(&user)
+	result := db.Save(&user)
+	if result.Error != nil {
+		return c.Status(fiber.StatusInternalServerError).JSON(fiber.Map{
+			"message": result.Error.Error(),
+		})
+	}
 
 	return c.JSON(fiber.Map{
-		"data": user,
+		"message": "User successfully updated",
+		"data": fiber.Map{
+			"id":    user.ID,
+			"name":  user.Name,
+			"email": user.Email,
+		},
 	})
 }
 
@@ -77,7 +129,16 @@ func DeleteUser(c *fiber.Ctx) error {
 	db := database.DB
 	var user model.User
 
-	db.First(&user, id)
+	parsedId, err := uuid.Parse(id)
+	if err != nil {
+		return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{"error": "Invalid UUID format"})
+	}
+
+	db.First(&user, parsedId)
+
+	if user.ID == uuid.Nil {
+		return c.Status(fiber.StatusNotFound).JSON(fiber.Map{"error": "No user found with given ID"})
+	}
 
 	db.Delete(&user)
